fix up engine tests
