+++ tests/python/pants_test/engine/test_mapper.py
@@ -13,32 +13,20 @@ from textwrap import dedent
 
 from pants.base.specs import DescendantAddresses, SiblingAddresses, SingleAddress
 from pants.build_graph.address import Address
-from pants.engine.addressable import BuildFileAddresses, SubclassesOf, addressable_list
+from pants.engine.addressable import BuildFileAddresses, Collection
 from pants.engine.build_files import UnhydratedStruct, create_graph_rules
 from pants.engine.engine import LocalSerialEngine
 from pants.engine.mapper import (AddressFamily, AddressMap, AddressMapper, DifferingFamiliesError,
                                  DuplicateNameError, UnaddressableObjectError)
 from pants.engine.nodes import Throw
 from pants.engine.parser import SymbolTable
-from pants.engine.selectors import SelectDependencies
-from pants.engine.struct import HasProducts, Struct
+from pants.engine.rules import TaskRule
+from pants.engine.selectors import Select, SelectDependencies
+from pants.engine.struct import Struct
 from pants.util.dirutil import safe_open
 from pants_test.engine.examples.parsers import JsonParser
 from pants_test.engine.scheduler_test_base import SchedulerTestBase
-
-
-class Target(Struct, HasProducts):
-  def __init__(self, name=None, configurations=None, **kwargs):
-    super(Target, self).__init__(name=name, **kwargs)
-    self.configurations = configurations
-
-  @property
-  def products(self):
-    return self.configurations
-
-  @addressable_list(SubclassesOf(Struct))
-  def configurations(self):
-    pass
+from pants_test.engine.util import Target, TargetTable
 
 
 class Thing(object):
@@ -221,10 +209,7 @@ class AddressFamilyTest(unittest.TestCase):
                                        {'one': Thing(name='one', age=37)})])
 
 
-class TargetTable(SymbolTable):
-  @classmethod
-  def table(cls):
-    return {'struct': Struct, 'target': Target}
+UnhydratedStructs = Collection.of(UnhydratedStruct)
 
 
 class AddressMapperTest(unittest.TestCase, SchedulerTestBase):
@@ -235,6 +220,15 @@ class AddressMapperTest(unittest.TestCase, SchedulerTestBase):
                                    parser_cls=JsonParser,
                                    build_patterns=('*.BUILD.json',))
     tasks = create_graph_rules(address_mapper, symbol_table_cls)
+    # TODO handle updating the rule graph when passed unexpected root selectors.
+    # Adding the following task allows us to get around the fact that SelectDependencies
+    # requests are not currently supported.
+    tasks.append(TaskRule(UnhydratedStructs,
+                          [SelectDependencies(UnhydratedStruct,
+                                              BuildFileAddresses,
+                                              field_types=(Address,),
+                                              field='addresses')],
+                          UnhydratedStructs))
 
     project_tree = self.mk_fs_tree(os.path.join(os.path.dirname(__file__), 'examples/mapper_test'))
     self.build_root = project_tree.build_root
@@ -247,7 +241,8 @@ class AddressMapperTest(unittest.TestCase, SchedulerTestBase):
                              type_alias='target')
 
   def resolve(self, spec):
-    select = SelectDependencies(UnhydratedStruct, BuildFileAddresses, field_types=(Address,), field='addresses')
+    # TODO reenable using SelectDependencies
+    select = Select(UnhydratedStructs)
     request = self.scheduler.selection_request([(select, spec)])
     result = LocalSerialEngine(self.scheduler).execute(request)
     if result.error:
@@ -257,7 +252,7 @@ class AddressMapperTest(unittest.TestCase, SchedulerTestBase):
     state, = result.root_products.values()
     if type(state) is Throw:
       raise Exception(state.exc)
-    return state.value
+    return state.value.dependencies
 
   def resolve_multi(self, spec):
     return {uhs.address: uhs.struct for uhs in self.resolve(spec)}
